+++ webapp/src/lib/qrcode/index.ts
@@ -7,14 +7,16 @@ import qrcode from 'qrcode-generator';
 type CredentialConfig = {
 	credential_configuration_ids: string[];
 	credential_issuer: string;
+	grants: {
+		authorization_code: {
+			authorization_server: string;
+		}
+	}
 };
 
 export function createIntentUrl(data: CredentialConfig): string {
-	const credentialConfigurationIds = encodeURIComponent(
-		data.credential_configuration_ids.join(',')
-	);
-	const credentialIssuer = encodeURIComponent(data.credential_issuer);
-	return `openid-credential-offer://?credential_configuration_ids=${credentialConfigurationIds}&credential_issuer=${credentialIssuer}`;
+	const credentialOffer = encodeURIComponent(JSON.stringify(data));
+	return `openid-credential-offer://?credential_offer=${credentialOffer}`;
 }
 
 export function generateQr(text: string, cellSize = 20) {